requirements.txt
beautifulsoup4==4.13.3
json2html==1.3.0
matplotlib==3.7.1
networkx==3.4.2
nltk==3.9.1
numpy==1.23.5
pandas==2.0.3
pdfkit==1.0.0
plotly==5.24.1
python-telegram-bot==21.10
PyYAML==6.0.2
requests==2.32.3
scipy==1.11.3
selenium==4.28.1
textblob==0.19.0
urllib3==2.3.0
kaleido==0.2.1
aiohttp==3.9.1
aiofiles==24.1.0